bundle-size: add tests for `app.js` (#18)
* Add es6 and jest environments to `.eslintrc`

* Fix comment description of `getBuildArtifactsFile`

* Move `db` object creation to top of file, due to how the test runner
initializes it inside the Probot app

* Ignore empty lines in OWNERS file

* Split the CLI runner and the setupDb function, to allow setting up the database for testing

* Update package.json with required packages for testing

* Avoid using setTimeout. Instead, when needing to retry, respond to the
caller with response.status(202).end(), and continue using a while loop

* Add tests for `bundle-size/app.js`

* Run tests as part of the PR check

* Run `npm ci` inside `bundle-size`

* Make test names more descriptive

* Split `const db` definition into two lines

+++ bundle-size/setup-db.js
@@ -14,29 +14,43 @@
  */
 'use strict';
 
+const {dbConnect} = require('./db');
 const log = require('fancy-log');
+
+/**
+ * Setup up the database schema.
+ *
+ * @param {knex} db database handler.
+ * @return {Promise<knex>} database handler.
+ */
+function setupDb(db) {
+  return db.schema.createTable('checks', table => {
+    table.string('head_sha', 40).primary();
+    table.string('base_sha', 40);
+    table.string('owner');
+    table.string('repo');
+    table.integer('pull_request_id');
+    table.integer('installation_id');
+    table.integer('check_run_id');
+    table.decimal('delta', 6, 2);
+  });
+}
+
+module.exports = {setupDb};
 
 /**
  * This file creates the database table that will be used by the GitHub App.
  *
  * Execute this file by running `npm run setup-db`. Make sure you set up the
- * environment variable first. See .env.example for details.
+ * database connection first. See db-config.example.js for details.
  */
-
-const db = require('./db').dbConnect();
-db.schema.createTable('checks', table => {
-  table.string('head_sha', 40).primary();
-  table.string('base_sha', 40);
-  table.string('owner');
-  table.string('repo');
-  table.integer('pull_request_id');
-  table.integer('installation_id');
-  table.integer('check_run_id');
-  table.decimal('delta', 6, 2);
-}).then(() => {
-  log.info('Database table `checks` created.');
-}).catch(error => {
-  log.error(error.message);
-}).then(() => {
-  return db.destroy();
-});
+if (require.main === module) {
+  const db = dbConnect();
+  setupDb(db).then(() => {
+    log.info('Database table `checks` created.');
+  }).catch(error => {
+    log.error(error.message);
+  }).then(() => {
+    db.destroy();
+  });
+}
